pyproject.toml
[tool.poetry]
name = "coffeeshops-search"
version = "0.1.0"
description = "Script that helps you to find nearest coffeeshop in Moscow"
authors = ["Lev Bazarnov <16255087+lbazarnov@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "coffeeshops_search"}]

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.28.1"
folium = "^0.13.0"
geopy = "^2.3.0"
flask = "^2.2.2"
python-dotenv = "^0.21.0"

[tool.poetry.scripts]
coffeeshops-search = "coffeeshops_search.main:main"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
